LC3 Assembly with Fibonacci Sequence
B1.asm repeatedly asks the user for a decimal digit and waits for the
user to type a digit ("0" to "9") followed by the return/enter key.
The program then prints the binary form of the digit as a 4 bit
unsigned value. Any value apart from a digit causes the program to
finish.

B2.asm iteratively prints out a list of Fibonacci numbers. The program
asks the user to enter a number between 3 and 23, and repeats until a
valid number is entered. If the value is n the program then prints out
the first n elements of the Fibonacci sequence and then stops.